chore: Remove dead code
